--- src/components/footer/Footer.scss
@@ -1,36 +1,0 @@
-$background-color-base: #121212;
-$bold-text-color: #fff;
-$base-text-color: #a7a7a7;
-$card-color: #242424;
-$border-radius-oval: 9999px;
-$font-family-base: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-
-.footer {
-  background-image: linear-gradient(90deg,#af2896,#509bf5);
-  color: $bold-text-color;
-  padding: 10px;
-  margin: 10px 0;
-}
-
-.footerLink {
-  color: $bold-text-color;
-  text-decoration: none;
-}
-
-.footerLeftHeader {
-  text-transform: uppercase;
-  letter-spacing: 0.1rem;
-  font-family: $font-family-base;
-  font-size: 0.75rem;
-}
-
-.footerLeftContent {
-  font-size: 1rem;
-}
-
-.footerSignUpBtn {
-  border-radius: $border-radius-oval !important;
-  font-weight: 600 !important;
-  width: 150px;
-  height: 50px;
-}
