+++ src/components/footer/FooterPreview.scss
@@ -1,0 +1,39 @@
+$background-color-base: #121212;
+$bold-text-color: #fff;
+$base-text-color: #a7a7a7;
+$card-color: #242424;
+$border-radius-oval: 9999px;
+$font-family-base: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+
+.preview {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background: linear-gradient(to right, #ae2896, #509bf5);
+  padding: 15px 40px;
+  display: flex;
+  justify-content: space-between;
+}
+.preview h6 {
+  color: #ffffff;
+  text-transform: uppercase;
+  font-weight: 400;
+  font-size: 12px;
+  margin-bottom: 10px;
+}
+.preview p {
+  color: #ffffff;
+  font-size: 14px;
+  font-weight: 500;
+}
+.preview button {
+  background-color: #ffffff;
+  color: #000000;
+  font-size: 16px;
+  font-weight: bold;
+  padding: 14px 30px;
+  border: 0px;
+  border-radius: 40px;
+  cursor: pointer;
+}
