+++ src/components/home-page/HomePage.scss
@@ -94,6 +94,7 @@ body {
   border: 0px;
   background-color: #090909;
   margin-right: 10px;
+  margin-top: 15px;
 }
 .topbar .navbar {
   display: flex;
@@ -156,9 +157,9 @@ body {
   overflow: hidden;
 }
 .spotify-playlists .list .item {
-  min-width: 140px;
-  width: 160px;
-  padding: 15px;
+  min-width: 160px;
+  width: 180px;
+  padding: 10px;
   background-color: #181818;
   border-radius: 6px;
   cursor: pointer;
@@ -206,37 +207,4 @@ body {
 .spotify-playlists hr {
   margin: 70px 0px 0px;
   border-color: #636363;
-}
-
-.preview {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background: linear-gradient(to right, #ae2896, #509bf5);
-  padding: 15px 40px;
-  display: flex;
-  justify-content: space-between;
-}
-.preview h6 {
-  color: #ffffff;
-  text-transform: uppercase;
-  font-weight: 400;
-  font-size: 12px;
-  margin-bottom: 10px;
-}
-.preview p {
-  color: #ffffff;
-  font-size: 14px;
-  font-weight: 500;
-}
-.preview button {
-  background-color: #ffffff;
-  color: #000000;
-  font-size: 16px;
-  font-weight: bold;
-  padding: 14px 30px;
-  border: 0px;
-  border-radius: 40px;
-  cursor: pointer;
 }
