+++ src/components/side-bar/SideBar.scss
@@ -10,7 +10,7 @@ $font-family-base: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
   left: 0;
   top: 0;
   bottom: 0;
-  width: 196px;
+  width: 250px;
   background-color: #000000;
   padding: 24px;
 }
